backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Financial Data & Analysis
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4
yfinance==0.2.35
pandas-datareader==0.10.0

# Schwab API
schwab-py==1.2.0

# Risk & Performance Calculations
empyrical==0.5.5
pyfolio==0.9.2

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# HTTP & API
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Data Visualization Support
plotly==5.18.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
